itinerary duration in integer seconds

+++ src/main/java/org/opentripplanner/api/resource/PlanGenerator.java
@@ -230,7 +230,7 @@ public class PlanGenerator {
         /* this is an affront to software development */
         fixupLegs(itinerary.legs, legsStates);
 
-        itinerary.duration = (double) lastState.getElapsedTimeSeconds();
+        itinerary.duration = lastState.getElapsedTimeSeconds();
         itinerary.startTime = makeCalendar(states[0]);
         itinerary.endTime = makeCalendar(lastState);
